requirements.txt
# base
django==2.2.12
djangorestframework==3.11.0
redis==3.4.1

# tests
pytest==5.4.1
pytest-django==3.9.0
pytest-cov==2.8.1
